+++ packages/vue-docgen-api/src/script-handlers/__tests__/mixinsHandler.ts
@@ -57,14 +57,17 @@ describe('mixinsHandler', () => {
 		const ast = babelParser().parse(src)
 		const path = resolveExportedComponent(ast).get('default')
 		if (path) {
-			await mixinsHandler(doc, path, ast, { filePath: '' })
+			await mixinsHandler(doc, path, ast, {
+				filePath: '',
+				validExtends: (fullFilePath: string) => !/[\\/]node_modules[\\/]/.test(fullFilePath)
+			})
 		}
 		expect(mockParse).toHaveBeenCalledWith(
-			doc,
 			expect.objectContaining({
 				filePath: './component/full/path',
 				nameFilter: ['default']
-			})
+			}),
+			doc
 		)
 		done()
 	})
@@ -82,13 +85,16 @@ describe('mixinsHandler', () => {
 			done.fail()
 			return
 		}
-		await mixinsHandler(doc, path, ast, { filePath: '' })
+		await mixinsHandler(doc, path, ast, {
+			filePath: '',
+			validExtends: (fullFilePath: string) => !/[\\/]node_modules[\\/]/.test(fullFilePath)
+		})
 		expect(mockParse).toHaveBeenCalledWith(
-			doc,
 			expect.objectContaining({
 				filePath: './component/full/path',
 				nameFilter: ['default']
-			})
+			}),
+			doc
 		)
 		done()
 	})
@@ -107,7 +113,10 @@ describe('mixinsHandler', () => {
 			return
 		}
 		mockResolvePathFrom.mockReturnValue('foo/node_modules/component/full/path')
-		await mixinsHandler(doc, path, ast, { filePath: '' })
+		await mixinsHandler(doc, path, ast, {
+			filePath: '',
+			validExtends: (fullFilePath: string) => !/[\\/]node_modules[\\/]/.test(fullFilePath)
+		})
 		expect(mockParse).not.toHaveBeenCalled()
 		done()
 	})
@@ -126,7 +135,10 @@ describe('mixinsHandler', () => {
 			done.fail()
 			return
 		}
-		await mixinsHandler(doc, path, ast, { filePath: '' })
+		await mixinsHandler(doc, path, ast, {
+			filePath: '',
+			validExtends: (fullFilePath: string) => !/[\\/]node_modules[\\/]/.test(fullFilePath)
+		})
 		expect(mockParse).not.toHaveBeenCalled()
 		done()
 	})