+++ packages/vue-docgen-api/src/script-handlers/extendsHandler.ts
@@ -37,18 +37,21 @@ export default async function extendsHandler(
 
 	// only look for documentation in the current project not in node_modules
 	const fullFilePath = pathResolver(extendsFilePath[extendsVariableName].filePath)
-	if (!/[\\/]node_modules[\\/]/.test(fullFilePath)) {
+	if (opt.validExtends(fullFilePath)) {
 		try {
 			const extendsVar = {
 				name: extendsFilePath[extendsVariableName].exportName,
 				path: fullFilePath
 			}
-			await parseFile(documentation, {
-				...opt,
-				filePath: fullFilePath,
-				nameFilter: [extendsFilePath[extendsVariableName].exportName],
-				extends: extendsVar
-			})
+			await parseFile(
+				{
+					...opt,
+					filePath: fullFilePath,
+					nameFilter: [extendsFilePath[extendsVariableName].exportName],
+					extends: extendsVar
+				},
+				documentation
+			)
 			extendsVar.name = documentation.get('displayName')
 		} catch (e) {
 			// eat the error