+++ packages/vue-docgen-api/tests/components/jsfile/MyButton.js
@@ -1,5 +1,8 @@
 export const name = 'EverButton'
 
+/**
+ * This is a button that represents a javascript only component, not a vue SFC
+ */
 export const Button = {
 	name,
 
@@ -46,6 +49,49 @@ export const Button = {
 
 		return h(
 			'div',
+			{
+				attrs: childAttrs,
+				on: this.$listeners
+			},
+			this.$slots.default
+		)
+	}
+}
+
+/**
+ * This is an input that represents another component extracted in the same file
+ */
+export const Input = {
+	name,
+
+	inheritAttrs: false,
+
+	props: {
+		variant: {
+			type: String,
+			default: 'solid'
+		},
+		variantColor: {
+			type: String,
+			default: 'gray'
+		},
+		size: {
+			type: [String, Number],
+			default: 'md'
+		},
+		isDisabled: {
+			type: Boolean
+		}
+	},
+
+	render(h) {
+		const childAttrs = {
+			disabled: this.isDisabled,
+			'aria-disabled': this.isDisabled
+		}
+
+		return h(
+			'input',
 			{
 				attrs: childAttrs,
 				on: this.$listeners